// src/models/Patient.ts
export interface Patient {
    _id: string;
    firstName: string;
    lastName: string;
    address: string;
    phone: string;
    gender: string;
    houseNumber: string;
    dateOfBirth: string;
    city: string;
    state: string;
    zipCode: string;
    medicalHistory: string
    medications: string;
}